fix: update imports for moved selection
